Major changes to support loading of different qa problems.
Unfortunately the dataloader does not support loading data of varying length. While the stories are padded, the questions are not, in the main architecture. Coincidentally within a problem the questions are all the same length, though. Therefore this was not noticed. In order to load multiple problem, the questions need to be padded. The models then have to work with the sorted packed sequences. The output has to be resorted to the original order when needed to combine questions and stories.